moved search bar to different file, added get requests to connect to back end
